Bug fixes
-- Fixed bugs in v6 and v7 filesystem interface code that prevented
seeking beyond the end of a file.

-- Fixed bug in the modernized version of dpadd for v6 that caused
various bad behaviors when writing files larger than 32767.

-- Relaxed access controls on setting the sticky bit to
match modern norms.

-- Other minor code clean-ups.

+++ ancient-src/v7/sys/fio.c
@@ -105,7 +105,7 @@ void
 openi(struct inode *ip, int16_t rw)
 {
 	dev_t dev;
-	register unsigned int maj;
+	register uint16_t maj;
 
 	dev = (dev_t)ip->i_un.i_rdev;
 	maj = major(dev);
